+++ models/index.js
@@ -1,8 +1,17 @@
 import productDao from './productDao';
 import commentDao from './commentDao';
-import userDao from './userDao';
 import cartDao from './cartDao';
 import likeDao from './likeDao';
+import loginDao from './loginDao';
+import signupDao from './signupDao';
 import listDao from './listDao';
 
-export { productDao, commentDao, cartDao, likeDao, userDao, listDao };
+export {
+  productDao,
+  commentDao,
+  cartDao,
+  likeDao,
+  loginDao,
+  signupDao,
+  listDao,
+};
