+++ package.json
@@ -23,7 +23,6 @@
     "bcrypt": "^5.0.1",
     "cookie-parser": "^1.4.5",
     "dotenv": "^10.0.0",
-    "eslint-plugin-import": "^2.25.2",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "prettier-plugin-prisma": "^3.1.1"
