+++ routes/commentRouter.js
@@ -3,7 +3,7 @@ import { commentController } from '../controllers';
 
 const router = express.Router();
 
-// router.post('/:id', commentController.addComment);
-// router.get('/:id', commentController.getCommentList);
+router.post('/:id', commentController.addComment);
+router.get('/:id', commentController.getCommentList);
 
 export default router;
