+++ routes/loginRouter.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import { loginController } from '../controllers';
+
+const router = express.Router();
+
+router.post('/', loginController.getUser);
+
+export default router;
