+++ services/index.js
@@ -2,7 +2,8 @@ import productServices from './productServices';
 import commentServices from './commentServices';
 import cartServices from './cartServices';
 import likeServices from './likeServices';
-import userServices from './userServices';
+import loginServices from './loginServices';
+import signupServices from './signupServices';
 import listServices from './listServices';
 
 export {
@@ -10,6 +11,7 @@ export {
   commentServices,
   cartServices,
   likeServices,
-  userServices,
+  loginServices,
+  signupServices,
   listServices,
 };
