+++ services/loginServices.js
@@ -1,14 +1,14 @@
-import { userDao } from '../models';
 import jwt from 'jsonwebtoken';
 import bcrypt from 'bcrypt';
 import dotenv from 'dotenv';
+import { loginDao } from '../models';
 import { ERROR } from '../utils/error';
 
 dotenv.config();
 const { secret } = process.env;
 
 const getUser = async (email, password) => {
-  const [user] = await userDao.getUser(email);
+  const [user] = await loginDao.getUser(email);
   if (user) {
     const checkPassword = await bcrypt.compare(password, user.password);
     if (checkPassword) {
@@ -22,25 +22,11 @@ const getUser = async (email, password) => {
 
 const getUserById = async id => {
   if (id) {
-    const userInfoById = await userDao.getUserById(id);
+    const userInfoById = await loginDao.getUserById(id);
     return userInfoById;
   } else {
     throw new error(ERROR.NO_USERS);
   }
 };
 
-const insertUser = async signupInfo => {
-  const { email, password, nickname } = signupInfo;
-  const [userInfo] = await userDao.getUsersByEmail(email);
-
-  if (!userInfo) {
-    const hash = await bcrypt.hash(password, 12);
-    return await userDao.insertUser(email, hash, nickname);
-  } else {
-    res
-      .status(500)
-      .send('이미 가입된 이메일입니다. 다른 이메일을 입력해주세요.');
-  }
-};
-
-export default { getUser, getUserById, insertUser };
+export default { getUser, getUserById };
